sdk: cleanup the deposit scripts
Modularize into 2 hardhat tasks, make them
easier to debug.
