+++ packages/sdk/tasks/index.ts
@@ -1,0 +1,2 @@
+import './deposit-eth'
+import './deposit-erc20'
